refactor(uselocale): update tests to follow new standards
